fix: fixed subscription submission
previously the submission of the form was impossible to complete, because the API don't have the CORS configurated for the frontend origin. In this way, to fix that I search and put the CORS inside candidates and email controllers. Configure axios baseURL and headers, toasts and async functions in the frontend also